// src/frontend/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

/**
 * Crawl
 */
export type Crawl = {
    id: number;
    website_url: string;
    url_filters?: Array<string> | null;
    pages: number;
    status: string;
    meta: {
        [key: string]: unknown;
    };
    created_at: string;
    updated_at: string;
    llms?: string | null;
    llms_full?: string | null;
};

/**
 * GetCrawl
 */
export type GetCrawl = {
    count: number;
    crawls: Array<Crawl>;
};

/**
 * PostCrawl
 */
export type PostCrawl = {
    website_url: string;
    gemini_api_key: string;
    url_filters?: Array<string> | null;
};

export type ApiCrawlListCrawlsData = {
    body?: never;
    path?: never;
    query?: {
        term?: string;
    };
    url: '/api/crawl';
};

export type ApiCrawlListCrawlsErrors = {
    /**
     * Validation Exception
     */
    400: {
        status_code: number;
        detail: string;
        extra?: null | {
            [key: string]: unknown;
        } | Array<unknown>;
    };
};

export type ApiCrawlListCrawlsError = ApiCrawlListCrawlsErrors[keyof ApiCrawlListCrawlsErrors];

export type ApiCrawlListCrawlsResponses = {
    /**
     * Request fulfilled, document follows
     */
    200: GetCrawl;
};

export type ApiCrawlListCrawlsResponse = ApiCrawlListCrawlsResponses[keyof ApiCrawlListCrawlsResponses];

export type ApiCrawlCreateCrawlData = {
    body: PostCrawl;
    path?: never;
    query?: never;
    url: '/api/crawl';
};

export type ApiCrawlCreateCrawlErrors = {
    /**
     * Validation Exception
     */
    400: {
        status_code: number;
        detail: string;
        extra?: null | {
            [key: string]: unknown;
        } | Array<unknown>;
    };
};

export type ApiCrawlCreateCrawlError = ApiCrawlCreateCrawlErrors[keyof ApiCrawlCreateCrawlErrors];

export type ApiCrawlCreateCrawlResponses = {
    /**
     * Document created, URL follows
     */
    201: Crawl;
};

export type ApiCrawlCreateCrawlResponse = ApiCrawlCreateCrawlResponses[keyof ApiCrawlCreateCrawlResponses];